% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invgamma.R
\name{dinvgamma}
\alias{dinvgamma}
\title{Density of the Inverse Gamma Distribution}
\usage{
dinvgamma(x, alpha, beta)
}
\arguments{
\item{x}{A numeric vector of the points at which the function is to be evaluated}

\item{alpha}{The alpha parameter of the inverse gamma distribution}

\item{beta}{The beta parameter of the inverse gamma distribution}
}
\value{
The density of an inverse-gamma variable at the points x
}
\description{
Vectorized density function for the inverse gamma distribution
}
\examples{
dinvgamma(seq(0.1, 100, 0.2), 3, 5)

}
