% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMCresult.R
\name{plot.MCMCresult}
\alias{plot.MCMCresult}
\title{Plots a MCMC Result from the IARREC package}
\usage{
\method{plot}{MCMCresult}(x, drop = 0, ...)
}
\arguments{
\item{x}{An MCMCresult S3 object}

\item{drop}{Indicates how much of the samples to ignore.
If it's a value between [0,1) then it ignores the first \eqn{drop * n} samples from
the result. If it's a value within [1, +Inf) then the first \eqn{drop} values are
ignored. Ignores nothing by default.}

\item{...}{Optional base R graphical parameters}
}
\description{
Plots a MCMC Result from the IARREC package
}
