% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.MCMCresult2}
\alias{plot.MCMCresult2}
\title{Plots a MCMC Result from the IARREC package}
\usage{
\method{plot}{MCMCresult2}(x, ...)
}
\arguments{
\item{x}{An MCMCresult S3 object}

\item{...}{Optional base R graphical parameters}
}
\description{
Plots a MCMC Result from the IARREC package
}
