% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slr_igig.R
\name{slr_igig}
\alias{slr_igig}
\title{Simple Linear Regression with Inverse Gamma Priors on the Non-Identifiable Parameters}
\usage{
slr_igig(numit, start, sigma, obs_Y, obs_X, beta0, beta1, mux, sx, st, sm)
}
\arguments{
\item{numit}{Number of iterations}

\item{start}{Vector with starting values for \eqn{\beta_0}, \eqn{\beta_1}, \eqn{\mu_x},
\eqn{\sigma_x^2}, \eqn{\sigma_r^2} and \eqn{\sigma_m^2} respectively}

\item{sigma}{The vector with the random walk parameter for each of the variables}

\item{obs_Y}{The vector of the Y observations}

\item{obs_X}{The vector of the X observations}

\item{beta0}{Mean and variance of the normal prior for \eqn{\beta_0}}

\item{beta1}{Mean and variance of the normal prior for \eqn{\beta_1}}

\item{mux}{Mean and variance of the normal prior for \eqn{\mu_x}}

\item{sx}{Parameters of the inverse gamma prior for \eqn{\sigma_x^2}}

\item{sm}{Parameters of the inverse gamma prior for \eqn{\sigma_m^2}}

\item{sr}{Parameters of the inverse gamma prior for \eqn{\sigma_r^2}}
}
\value{
An 'MCMCresultSLR' object.
}
\description{
Runs a MCMC to generate a sample for the posterior of the parameters of a simple
linear regression model with errors in the covariates.
}
\examples{
x <- rnorm(60, 50, sqrt(20)/2)
xs <- x + rnorm(60, 0, sqrt(20)/2)
y <- 50 + 50*x + rnorm(60, 0, sqrt(50))
O <- slr_igig(75000, rep(1, 6), rep(0.6, 6), y, xs, c(0, 500), c(0, 500), c(0, 500), c(3, 20), c(2, 50), c(3, 20))

}
