% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.MCMCresultSLR}
\alias{plot.MCMCresultSLR}
\title{Plots a MCMC Result from the IARREC package}
\usage{
\method{plot}{MCMCresultSLR}(x, pause = TRUE, ...)
}
\arguments{
\item{x}{An MCMCresult S3 object}

\item{pause}{Indicates if should pause in between plots. Defaults to TRUE.}

\item{...}{Optional base R graphical parameters}
}
\description{
Plots a MCMC Result from the IARREC package
}
